# server/swagger.yaml
openapi: 3.0.0
info:
  title: SchoolMS
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /subject/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Subject
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: subjectname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Subject
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: subjectname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    post:
      summary: CRUD ACTION update
      tags:
        - Subject
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: subjectname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Subject"
...
  /subject/{id}/get_Class:
    get:
      summary: CRUD ACTION get_Class
      tags:
        - Subject
      parameters:
        - name: class
          in: path
          description: 
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
  /subject/{id}/get_Teacher:
    get:
      summary: CRUD ACTION get_Teacher
      tags:
        - Subject
      parameters:
        - name: teachername
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  Teacher  -----
...
  /teacher/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Teacher
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: teachername
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Teacher
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: teachername
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    post:
      summary: CRUD ACTION update
      tags:
        - Teacher
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
        - name: teachername
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teacher"
...
  /teacher/{id}/get_Class:
    get:
      summary: CRUD ACTION get_Class
      tags:
        - Teacher
      parameters:
        - name: class
          in: path
          description: 
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
  /teacher/{id}/get_Exam:
    get:
      summary: CRUD ACTION get_Exam
      tags:
        - Teacher
      parameters:
        - name: examname
          in: path
          description: 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  Test  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Subject:
      properties:
        _id:
          type: string
        averageresultinlastexam:
          type: number
        class:
          type: number
        period:
          type: number
        subjectname:
          type: string
        teachername:
          type: string
      required:
        -  _id
        - class
        - subjectname
        - teachername
    Teacher:
      properties:
        _id:
          type: string
        class:
          type: number
        teachername:
          type: string
      required:
        -  _id
        - class
        - teachername
...